check for file locking in batch building
Added the while loop and try block to check for file locking in batch builds. In batch builds in Visual Studio the pre-build sometimes tries to run before a parallel build has unlocked the files, causing an error.  This loop will cause the pre-build to pause until the destination files are unlocked and can be overwritten.